McLib now pre-generates the dataset.
You no longer have to pre-calculate your observable for determining
the correlation length, greatly simplifying a lot of code.
You just need to provide your estimator, and what your estimator
requires; from there, it will generate the observable with the estimator
in the way you wanted it slice.  Should be more correct, robust, and
result in less code.

w_reweight is using this, but w_direct is not.  mclib will also require
more changes to 'purge' putting in the pre-calculated dataset.