Fixed start at iter 1 bug w/ w_reweight.
The function accumulate_statistics did not work properly when starting
at any iteration than the first.  The current fix is to simply always
accumulate data starting from the beginning; the insert array must have
the iteration matched with the index (that is, iteration 240 must be
able to use its iteration value to properly index the array).