Fixed scaling for trans matrix, error in mclib
All values output from the estimator for w_reweight are now properly
scaled by the subtau value, meaning that results are stored per tau, as
intended.

In addition, the mclib now checks to see if he pre_calculated dataset
has a length of 0; if this is the case, something has gone wrong with
the estimator and must be fixed.  That is, when returning values, we
shouldn't have to catch NaNs at the mclib level.  The estimator function
should handle unknown values, as only it can decide on an appropriate
'default' return value.