Minor bugfix: Changed check for seg_id defined
The previous behavior checked `seg_ids == None`, which raises an error if `seg_ids` is defined. This was addressed by wrapping the whole thing in a try/except block.

However, the intended behavior can be produced by just checking with `is` instead of `==`, which correctly evaluates to `False` if `seg_ids` is defined instead of erroring out. Although the contents of the try block are fairly simple, doing the evaluation this way without the unnecessary try block prevents other potential errors in the block from being incorrectly ignored.

Former-commit-id: 6060e4b07564c027ec582e75630523736b8864f1 [formerly 56c3873a01212c1ad94e551e0911cc3852cf9baa]
Former-commit-id: f0ce51e01731b4b5872528ce7eda49ef9b05fe42